backend-django/requirements.txt
#django-filter
#django-webpack-loader
#git+https://github.com/intelie/django-oidc-auth.git
#markdown
#requests
bcrypt
beautifulsoup4
dataclasses_json
deepdiff
django
django-allauth
django-cleanup
django-environ
django-oidc-auth
djangorestframework
djangorestframework-simplejwt
drf-oidc-auth
drf-spectacular
lxml
mysql-connector-python
mysqlclient
newspaper3k
pillow
# pyeuropeana
pymongo
pyyaml
simplejwt
six
torch
trafilatura
transformers
uritemplate
